# internal/services/identity_service/docs/v1/swagger.yaml
definitions:
  AuthenticateRequest:
    properties:
      password:
        type: string
      usernameOrEmailAddress:
        type: string
    required:
    - password
    - usernameOrEmailAddress
    type: object
  AuthenticateResult:
    properties:
      access_token:
        type: string
      expire_in_seconds:
        type: integer
      refresh_token:
        type: string
      refresh_token_expire_in_seconds:
        type: integer
    type: object
  CreateOrEditRoleDto:
    properties:
      grantedPermissions:
        items:
          type: string
        type: array
      id:
        type: integer
      name:
        type: string
    required:
    - name
    type: object
...
  CreateRoleDto:
    properties:
      grantedPermissions:
        items:
          type: string
        type: array
      id:
        type: integer
      name:
        type: string
    required:
    - name
    type: object
  CreateUserDto:
    properties:
      email:
        maxLength: 256
        type: string
      firstName:
        maxLength: 64
        minLength: 3
        type: string
      id:
        type: integer
      lastName:
        maxLength: 64
        minLength: 3
        type: string
      password:
        type: string
      roleIds:
        items:
          type: integer
        type: array
      userName:
        maxLength: 256
        minLength: 8
        type: string
    required:
    - email
    - firstName
    - lastName
    - userName
    type: object
  EditRoleDto:
    properties:
      grantedPermissions:
        items:
          type: string
        type: array
      id:
        type: integer
      name:
        type: string
    required:
    - name
    type: object
  EditUserDto:
    properties:
      email:
        maxLength: 256
        type: string
      firstName:
        maxLength: 64
        minLength: 3
        type: string
      id:
        type: integer
      lastName:
        maxLength: 64
        minLength: 3
        type: string
      password:
        type: string
      roleIds:
        items:
          type: integer
        type: array
      userName:
        maxLength: 256
        minLength: 8
        type: string
    required:
    - email
    - firstName
    - lastName
    - userName
    type: object
  GetAllPermissionResult:
    properties:
      items:
        items:
          $ref: '#/definitions/PermissionGroupDto'
        type: array
    type: object
...
  GetRoleByIdResult:
    properties:
      role:
        $ref: '#/definitions/CreateOrEditRoleDto'
    type: object
  GetRolesResult:
    properties:
      items:
        items:
          $ref: '#/definitions/RoleDto'
        type: array
      totalCount:
        type: integer
    type: object
...
  OAuthAuthenticateRequest:
    properties:
      code:
        type: string
      redirect_uri:
        type: string
    required:
    - code
    - redirect_uri
    type: object
  PermissionDto:
    properties:
      displayName:
        type: string
      isGranted:
        type: boolean
      name:
        type: string
    type: object
  PermissionGroupDto:
    properties:
      groupName:
        type: string
      permissions:
        items:
          $ref: '#/definitions/PermissionDto'
        type: array
    type: object
  RefreshTokenRequest:
    properties:
      token:
        minLength: 10
        type: string
    required:
    - token
    type: object
  RefreshTokenResult:
    properties:
      accessToken:
        type: string
      expireInSeconds:
        type: integer
    type: object
  RoleDto:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  UpdateUserPermissionDto:
    properties:
      grantedPermissions:
        items:
          type: string
        type: array
    type: object
  UserDto:
    properties:
      email:
        type: string
      firstName:
        type: string
      id:
        type: integer
      lastName:
        type: string
      userName:
        type: string
    type: object
...
  UserPermissionsResult:
    properties:
      items:
        items:
          type: string
        type: array
    type: object
info:
  contact: {}
paths:
  /api/v1/identities/app-permissions:
    get:
      consumes:
      - application/json
      description: Get All App Permissions
      operationId: GetAllAppPermissions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetAllPermissionResult'
      security:
      - ApiKeyAuth: []
      summary: Get All App Permissions
      tags:
      - Identities
  /api/v1/identities/authenticate:
    post:
      consumes:
      - application/json
      description: Authenticate
      operationId: Authenticate
      parameters:
      - description: AuthenticateRequest
        in: body
        name: AuthenticateRequest
        required: true
        schema:
          $ref: '#/definitions/AuthenticateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/AuthenticateResult'
      security:
      - ApiKeyAuth: []
      summary: Authenticate
      tags:
      - Identities
...
  /api/v1/identities/oauth-authenticate:
    post:
      consumes:
      - application/json
      description: OAuthAuthenticate
      operationId: OAuthAuthenticate
      parameters:
      - description: OAuthAuthenticateRequest
        in: body
        name: OAuthAuthenticateRequest
        required: true
        schema:
          $ref: '#/definitions/OAuthAuthenticateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/AuthenticateResult'
      security:
      - ApiKeyAuth: []
      summary: OAuthAuthenticate
      tags:
      - Identities
  /api/v1/identities/refresh-token:
    post:
      consumes:
      - application/json
      description: Refresh access token
      operationId: RefreshToken
      parameters:
      - description: RefreshTokenRequest
        in: body
        name: RefreshTokenRequest
        required: true
        schema:
          $ref: '#/definitions/RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RefreshTokenResult'
      security:
      - ApiKeyAuth: []
      summary: Refresh access token
      tags:
      - Identities
  /api/v1/identities/role:
    post:
      consumes:
      - application/json
      description: Create new role
      operationId: CreateRole
      parameters:
      - description: CreateRoleDto
        in: body
        name: CreateRoleDto
        schema:
          $ref: '#/definitions/CreateRoleDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RoleDto'
      security:
      - ApiKeyAuth: []
      summary: Create new role
      tags:
      - Roles
    put:
      consumes:
      - application/json
      description: Update role
      operationId: UpdateRole
      parameters:
      - description: EditRoleDto
        in: body
        name: EditRoleDto
        schema:
          $ref: '#/definitions/EditRoleDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/RoleDto'
      security:
      - ApiKeyAuth: []
      summary: Update role
      tags:
      - Roles
  /api/v1/identities/role/{roleId}:
    delete:
      consumes:
      - application/json
      description: Delete role
      operationId: DeleteRole
      parameters:
      - description: Role Id
        in: path
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Delete role
      tags:
      - Roles
    get:
      consumes:
      - application/json
      description: Get role by id
      operationId: GetRoleById
      parameters:
      - description: Role Id
        in: path
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetRoleByIdResult'
      security:
      - ApiKeyAuth: []
      summary: Get role by id
      tags:
      - Roles
  /api/v1/identities/roles:
    get:
      consumes:
      - application/json
      description: Get all roles
      operationId: GetAllRoles
      parameters:
      - in: query
        name: filters
        type: string
      - in: query
        name: maxResultCount
        type: integer
      - in: query
        name: skipCount
        type: integer
      - in: query
        name: sorting
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetRolesResult'
      security:
      - ApiKeyAuth: []
      summary: Get all roles
      tags:
      - Roles
...
  /api/v1/identities/user:
    post:
      consumes:
      - application/json
      description: Create new user
      operationId: CreateUser
      parameters:
      - description: CreateUserDto
        in: body
        name: CreateUserDto
        schema:
          $ref: '#/definitions/CreateUserDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UserDto'
      security:
      - ApiKeyAuth: []
      summary: Create new user
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update user
      operationId: UpdateUser
      parameters:
      - description: EditUserDto
        in: body
        name: EditUserDto
        schema:
          $ref: '#/definitions/EditUserDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UserDto'
      security:
      - ApiKeyAuth: []
      summary: Update user
      tags:
      - Users
...
  /api/v1/identities/user/{userId}/permissions:
    get:
      consumes:
      - application/json
      description: Get user permissions
      operationId: GetUserPermissions
      parameters:
      - description: User Id
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/UserPermissionsResult'
      security:
      - ApiKeyAuth: []
      summary: Get user permissions
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update user permissions
      operationId: UpdateUserPermissions
      parameters:
      - description: User Id
        in: path
        name: userId
        required: true
        type: integer
      - description: UpdateUserPermissionDto
        in: body
        name: UpdateUserPermissionDto
        schema:
          $ref: '#/definitions/UpdateUserPermissionDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update user permissions
      tags:
      - Users
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"